Enhance JSON schemas (#320)
* Add transformation layer between integration models and API calls for data that is persisted
* Add specs for a bunch of API transformation things
* Remove branch-naming hacks
* Remove deprecated method `normalize` and its usage
* Refactor BuildArtifacts Stimulus controller to be generic
* Add data migrations to pertain to new schema
* Add new build error status for Bitrise
* Add a validation on max build number accepted by android

+++ db/data/20221005110149_cleanup_to_avoid_carefully_migrating.rb
@@ -1,16 +1,20 @@
 class CleanupToAvoidCarefullyMigrating < ActiveRecord::Migration[7.0]
-  # there are no active users at this point so this is safe
   def up
-    Releases::CommitListener.delete_all
-    BuildArtifact.delete_all
-    Releases::Step::Run.delete_all
-    SignOff.delete_all
-    Releases::Commit.delete_all
-    Releases::Step::Run.delete_all
-    Releases::Step.delete_all
-    Releases::Train::Run.delete_all
-    TrainSignOffGroup.delete_all
-    Releases::Train.delete_all
+    return unless Rails.env.production?
+
+    # there are no active users at this point so this is safe
+    ActiveRecord::Base.transaction do
+      Releases::CommitListener.delete_all
+      BuildArtifact.delete_all
+      Releases::Step::Run.delete_all
+      SignOff.delete_all
+      Releases::Commit.delete_all
+      Releases::Step::Run.delete_all
+      Releases::Step.delete_all
+      Releases::Train::Run.delete_all
+      TrainSignOffGroup.delete_all
+      Releases::Train.delete_all
+    end
   end
 
   def down
